+++ lib/constants.dart
@@ -3,5 +3,5 @@ const kDefaultSelectedBorderWidth = 1.0;
 const kDefaultPadding = 16.0;
 const kDefaultChipSpacing = 8.0;
 const kDefaultChipRadius = 30.0;
-const kMinDropSelectorWidth = 250.0;
+const kMaxDropSelectorWidth = 250.0;
 const String kPackageName = 'lite_forms';